refact
Move logic out of nickname_combobox and overlay into Nickname mgr

Add ssort to project
